refactor: 비대면 종료 시 요약 생성하고 요약 조회에서 생성된 요약을 조회하도록 수정 (#82)
