+++ src/main/java/npng/handdoc/telemed/controller/TelemedController.java
@@ -34,7 +34,7 @@ public class TelemedController {
         return ResponseEntity.ok(ApiResponse.from(telemedService.join(userDetails.getId(), reservationId)));
     }
 
-    @Operation(summary = "(환자, 의사) 진료실 종료 API", description = "진료 종료 버튼을 통해 화상 통화를 마무리합니다. roomId를 입력해주세요.")
+    @Operation(summary = "(환자, 의사) 진료실 종료 API", description = "진료 종료 버튼을 통해 화상 통화를 마무리하고 요약을 생성하여 db에 넣습니다. roomId를 입력해주세요.")
     @PostMapping("/{roomId}/end")
     public ResponseEntity<ApiResponse<Object>> end(@AuthenticationPrincipal CustomUserDetails userDetails,
                                                    @PathVariable String roomId) {
